Move query builder logic to trait
